requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.2
django-cors-headers>=4.3.0
django-ratelimit>=4.0.0
openpyxl>=3.1.0
pillow>=10.0.0
python-decouple>=3.8 